complete video build only if construction files need cleaning

+++ video_service.py
@@ -64,18 +64,19 @@ def commit_to_db(media_path):
 
 def complete_video_build(media_path, result):
     ts_files = [f for f in glob.glob(media_path+'/*.ts')]
-    sorted_ts_files = sorted(ts_files, key=os.path.getctime, reverse=True)
-    latest_datetime = datetime.strptime(sorted_ts_files[0][len(media_path)+1:-3], '%Y-%m-%d_%H-%M-%S')
-    fifteen_after = datetime.combine(date.today(), result.end_time) + timedelta(minutes=15)
-    fifteen_before = datetime.combine(date.today(), result.start_time) + timedelta(minutes=15)
-    if latest_datetime < fifteen_before or latest_datetime > fifteen_after:
-        try:
-            os.remove(media_path+'/vidlist.txt')
-            for vid in glob.glob(media_path+'/*.ts'):
-                os.remove(vid)
-                print "removed: %s" % (vid)
-        except:
-            print "files couldn't be removed, we'll have to try this again."
+    if ts_files:
+        sorted_ts_files = sorted(ts_files, key=os.path.getctime, reverse=True)
+        latest_datetime = datetime.strptime(sorted_ts_files[0][len(media_path)+1:-3], '%Y-%m-%d_%H-%M-%S')
+        fifteen_after = datetime.combine(date.today(), result.end_time) + timedelta(minutes=15)
+        fifteen_before = datetime.combine(date.today(), result.start_time) + timedelta(minutes=15)
+        if latest_datetime < fifteen_before or latest_datetime > fifteen_after:
+            try:
+                os.remove(media_path+'/vidlist.txt')
+                for vid in glob.glob(media_path+'/*.ts'):
+                    os.remove(vid)
+                    print "removed: %s" % (vid)
+            except:
+                print "files couldn't be removed, we'll have to try this again."
 
 
 today_string = datetime.strftime(datetime.now(), '%Y_%m_%d')
